Fix crash when setting RTCRtpEncodingParameter.maxBitrate to a non-positive value.
This will result in using a default value instead.

The crash is caused by configuring audio send stream with a non-positive max_bitrate_bps and it expects a positive value (or special value -1).

Also remove support for setting min bitrate since it is not in the spec and there are no tests.

Bug: chromium:1377286
Change-Id: I19af29bff79eda5c0fbb1e528fced7976f5c2511
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/284640
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Jakob Ivarsson‎ <jakobi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#38719}