Remove global gtest_parallel timeout.
It is causing flakiness on win asan bots rtc_unittests.
The gtest_parallel timeout is set to 15 minutes but there is already a swarming timeout set to one hour.
https://luci-analysis.appspot.com/p/webrtc/clusters/testname-v3/218550890147469e975d34b17d4c6545#recent-failures
https://luci-analysis.appspot.com/p/webrtc/clusters/testname-v3/968d49e7d00955cc65fa84dfa54b7c2a#recent-failures

There is a comment stating that it's useful because it gets information when the swarming timeout is reached but:
* swarming timeouts are more likely to occur on device phone before impacting desktop bots.
* if a swarming timeout occurs, isolating which tests take too long shouldn't be too hard.

Change-Id: I2089918e8cd266510d58bac8c91d92b6c8e81728
Bug: webrtc:14790
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/288384
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Commit-Queue: Jeremy Leconte <jleconte@google.com>
Cr-Commit-Position: refs/heads/main@{#38918}