Use //third_party/cpu_features directly
The Android SDK has a target that references the NDK, which references
the cpu_features library. Directly take a dependency on the final
library, as all the targets in the chain are just soft-transition
targets. This also enables us to remove the Android NDK dependency from
the BUILD.gn file of the SDK.

Bug: chromium:1446042
Test: CQ
Change-Id: I5178faa71a62711e5d96af71c0715acdc1d366db
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/307380
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Auto-Submit: Prashanth Swaminathan <prashanthsw@google.com>
Commit-Queue: Mirko Bonadei <mbonadei@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#40200}