Migrate Android NDK to CIPD [1/2]
The Android NDK dependency is moving to a CIPD bucket to reduce the
checkout cost and to eventually move to NDK v25. This introduces the
NDK into an 'android_toolchain' directory. Following the roll of
chromium/base in this repository, a second change will delete the old
'android_ndk' checkout. As a result, the checkout size of this
repository will temporarily increase.

This change also updates the license generation scripts and build
helpers to ensure the android_toolchain is also accounted for.

Bug: chromium:1448383
Test: Verified local builds of WebRTC.
Change-Id: Ibe667be241e5a454d884482061dd10b9850be25f
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/307540
Commit-Queue: Prashanth Swaminathan <prashanthsw@google.com>
Reviewed-by: Christoffer Jansson <jansson@webrtc.org>
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#40239}