kDummyAudio now also creates Dummy ADM on Android
The old Android ADM was removed in https://webrtc-review.googlesource.com/c/src/+/271841.

This change resulted in a NULL as result when asking for a
kDummyAudio ADM on Android.

The small change below should ensure that a dummy ADM can be
created on Android as well.

Bug: webrtc:7452, b/291275589
Change-Id: I2c995ce6ba9a4117e3e39596546b133fe1c49204
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/311946
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Commit-Queue: Henrik Andreassson <henrika@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#40440}