Each spatial layer can only have 1 QP value.
As the code is now, it looks like it accepts that a spatial layer can have more than 1 QP value. These QP values according to the code are summed. However, to my best knowledge this cannot be the case and makes the code hard to read. Therefore, updating it with a check such that it checks that each spatial layer only have 1 QP value and would be easier for a future code reader.

Bug: webrtc:357636606
Change-Id: I650cac724811a1ddc7ab8933c1e1ac5fe844b61c
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/358101
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Reviewed-by: Artem Titov <titovartem@webrtc.org>
Commit-Queue: Emil Vardar (xWF) <vardar@google.com>
Cr-Commit-Position: refs/heads/main@{#42736}