Document when the dependency descriptor can be negotiated but not sent
This can happen when VP8 simulcast is negotiated while two-byte header
extensions are not negotiated via extmap-allow-mixed. For VP8 the
DD extension would be 23 bytes long which exceeds the maximum size
of 15 bytes for a one-byte header extension.

To test, revert
  https://github.com/fippo/simulcast-playground/commit/f04b52b4a782addc97fd0a74923cda30122f4e78
and test using VP8.

Note that this works for VP9, AV1, H264 out of the box.

BUG=webrtc:40191093

Change-Id: I2f5d04d8b58b71d32547b06fab6b9a9006df9f1a
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/359623
Commit-Queue: Philipp Hancke <phancke@meta.com>
Reviewed-by: Danil Chapovalov <danilchap@webrtc.org>
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#42786}