Grant CQ+2 access to project-webrtc-submit-access.
The reason we are doing this is because the "committers" group grants access to Code-Review, and we now have a policy that such groups must only contain direct human users. However, we still want robot accounts to be able to CQ+2, so we now have the -submit-access group that includes anyone who can CQ+2, not just committers.

Bug: chromium:356276684
Change-Id: I39c3b92110ff10f53a2825f5e12690580331364d
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/363480
Reviewed-by: Christoffer Dewerin <jansson@webrtc.org>
Auto-Submit: Joey Scarr <jsca@google.com>
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Commit-Queue: Mirko Bonadei <mbonadei@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#43073}