Add support for timestamps to TransformableVideoFrameInterface
This covers receive time and the timestamps associated with the
Absolute Capture Time header extension.

Since TransformableAudioFrameInterface already had support, this CL
moves the methods to the parent TransformableFrameInterface.

This CL only intends to change behavior for receiver frames.

The change in RTPSenderVideoFrameTransformerDelegate is basically a
rename of an existing GetCaptureTime() method to the new
CaptureTime(), with no behavior change.


Bug: chromium:391114797
Change-Id: I3d0bc55287e5bbc3ffcf80582e70a232589cd4ce
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/378920
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Guido Urdaneta <guidou@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#43998}