Add field trial that change g2g metric to use abs. capture time.
G2G (glass to glass) latency metric uses RTP timestamps to estimate the metric. The RTP timestamps are converted to senders NTP time using data from Sender Reports. This work fine in most cases, but not when Virtual Video SSRCs are used.

This field trail changes the G2G metric to use the absolute capture time header extension to get senders NTP timestamp. This is designed to work with Virtual SSRCs, and is currently used for audio end to end latency metric.

Enables this should improve G2G latency metric calculations with VVSSRCs , and create the same results in other cases. It should not create any other side effects. The feature is hidden behind the flag so we can make sure that's the case as we roll this out.

Bug: webrtc:401512883
Change-Id: I7a75d0a430f1fd1bcd79d9a228a7429300d5fafe
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/380421
Reviewed-by: Åsa Persson <asapersson@webrtc.org>
Reviewed-by: Philip Eliasson <philipel@webrtc.org>
Commit-Queue: Olov Brändström <brandstrom@google.com>
Cr-Commit-Position: refs/heads/main@{#44126}