Revert "fix: h26x packet buffer video artifacts"
In the case of H264 the fix to avoid incomplete keyframes being
assembled is to add `sps-pps-idr-is-keyframe` as an fmtp parameter
during negotiation.

This CL introduces a bug where SEI NALUs will block an actually complete
frame from being assembled as a SEI NALU is not a
`is_first_packet_in_frame`.

This reverts commit eafee5e3d68a7356dd415d2b3eee53017140d38b.

Bug: chromium:402547556, webrtc:384391181, chromium:402520966
Change-Id: Id3a3731b1d4171dff19ee4f467fe5c50e90bc7be
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/380860
Reviewed-by: Sergey Silkin <ssilkin@webrtc.org>
Reviewed-by: Henrik Boström <hbos@webrtc.org>
Commit-Queue: Philip Eliasson <philipel@webrtc.org>
Reviewed-by: Jianlin Qiu <jianlin.qiu@intel.com>
Cr-Commit-Position: refs/heads/main@{#44108}